[#395] Additional WIP
