+++ Client/src/Components/UserReview/UserReviewList.tsx
@@ -1,42 +1,40 @@
 import { Box, Button, Grid, Link, TextField, Typography } from '@material-ui/core';
 
 import { IApprovalDatasetModel } from '../../../../Server/src/models/interfaces/DatasetModelInterface';
-import { IApprovedDatasetModel } from '../../Models/Datasets/IApprovedDatasetModel';
+import { IApprovedDatasetModel, IFlaggedDatasetQuery } from '../../Models/Datasets/IApprovedDatasetModel';
 import React from 'react';
 import { classStyles } from '../../appTheme';
 import { DatasetViewModal } from '../DatasetUpload/DatasetViewModal';
 import { DatasetUploadForm } from '../DatasetUpload/DatasetUploadForm';
 import { useState } from 'react';
 import { post } from '../../Remote/FluentRequest';
+import { DatasetModal } from './DatasetModal';
+import { adminApprovedDataset, callRejectDataset } from '../../Remote/Endpoints/DatasetEndpoint';
+import { UserReviewRow } from './UserReviewRow';
+
+
 
 interface IProps {
-    userDatasets: IApprovedDatasetModel[],
-    handleModal: () => void
+    userDatasets: IApprovedDatasetModel[]
 }
 
 export const UserReviewList = (props: IProps) => {
 
-    const { userDatasets, handleModal } = { ...props }
+    const { userDatasets } = { ...props }
 
     return (
         <Grid container justify="center" spacing={3}>
             <Grid item xs={3}>
                 <Box className={classStyles().datasetBorder}>
-                    <br></br>
                     {userDatasets && userDatasets.map(dataset => {
                         return (
-                            <Box>
-                                <Grid item>
-                                    <Button size="small" id="view-dataset" onClick={handleModal} color="primary" variant="contained">{dataset.dataset_name} </Button>
-                                </Grid>
-                                <Grid item>
-                                    <p> {dataset.datasetFlaggedComment} </p>
-                                </Grid>
-                            </Box>
+                            <UserReviewRow
+                                dataset={dataset}
+                            />
                         )
                     })}
                 </Box>
             </Grid>
-        </Grid >
+        </Grid>
     )
 }
